feat: set up the project
