+++ README.md
@@ -1,4 +1,3 @@
 # spring-gift-order
-
-# step 0
-- 상품 고도화 코드 옮겨오기
+step0
+[x] 상품 고도화 코드 옮기기
