+++ src/main/java/gift/controller/PageController.java
@@ -15,5 +15,4 @@ public class PageController {
     public String wishlistPage() {
         return "wishlist";
     }
-
 }
