+++ src/main/java/gift/dto/ProductRequest.java
@@ -2,10 +2,12 @@ package gift.dto;
 
 import gift.domain.ProductOption;
 import gift.validation.NoKakao;
-import jakarta.validation.constraints.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Positive;
+import jakarta.validation.constraints.Size;
 
 import java.util.List;
-
 public class ProductRequest {
     private Long id;
