+++ src/main/resources/templates/products/list.html
@@ -99,4 +99,3 @@
 </body>
 </html>
 
-
